Adjust auto off
Turns out my laptop still pulls 32W when it's fully charged. I know the
toaster uses hundreds of watts. I don't really use anything else on the
inverter right now so this is fine. I could try to detect the previous
load and see if I think nothing is using it, but for now this is fine.